# server/OpenAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie API",
    "description": "API for users to access accounts, profiles, and group interactions.",
    "contact": { "email": "your-email@example.com" },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://movieapp-backend-actions-gxfcdkh7h6hccvfu.northeurope-01.azurewebsites.net",
      "description": "azure development server"
    },
    {
      "url": "http://localhost:{port}/",
      "description": "Local development server",
      "variables": {
        "port": {
          "description": "Port on which the server is running",
          "default": 3000
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/group/": {
      "get": {
        "tags": ["groups"],
        "summary": "Retrieve all groups",
        "description": "Fetch a list of all groups.",
        "operationId": "getAllGroupsListing",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Groups list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/group_response"
                  },
                  "example": [
                    {
                      "id": 1,
                      "name": "Developers Group",
                      "description": "A group for developers",
                      "creator_id": 101
                    },
                    {
                      "id": 2,
                      "name": "Designers Hub",
                      "description": "A group for designers",
                      "creator_id": 102
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/all": {
      "get": {
        "tags": ["groups"],
        "summary": "Retrieve all groups accessible by the authenticated user",
        "description": "Fetch a list of all groups where the authenticated user is a member.",
        "operationId": "getGroupsByAuth",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Groups list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/group_detail_response"
                  },
                  "example": [
                    {
                      "id": 1,
                      "name": "Developers Group",
                      "creator_id": 201,
                      "member_count": 25
                    },
                    {
                      "id": 2,
                      "name": "Designers Hub",
                      "creator_id": 202,
                      "member_count": 15
                    }
                  ]
                }
              }
            }
          },
          "400": { "description": "Bad Request - Invalid user ID provided" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/userGroup/{id}": {
      "get": {
        "tags": ["groups"],
        "summary": "Retrieve groups associated with a user by the ID from url",
        "description": "Fetch a list of groups where the user, specified by their ID, is a member. Includes group details and member count.",
        "operationId": "getGroupsByUrlId",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the user whose groups are being retrieved",
            "schema": {
              "type": "integer",
              "example": 101
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groups list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/group_detail_response"
                }
              }
            }
          },
          "400": { "description": "Bad Request - Invalid user ID provided" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/create": {
      "post": {
        "tags": ["groups"],
        "summary": "Create a new group",
        "description": "Allows an authenticated user to create a new group with a specified name and description. The authenticated user is set as the group's creator and automatically added as a member.",
        "operationId": "createNewGroup",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_group_body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/create_group_response"
                }
              }
            }
          },
          "400": { "description": "Bad Request - Invalid group name" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/{groupId}": {
      "get": {
        "tags": ["groups"],
        "summary": "Get Group by Group ID",
        "description": "Retrieve details of a specific group by its unique ID.",
        "operationId": "getGroupDetails",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_group_response"
                }
              }
            }
          },
          "404": { "description": "Group not found" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      },
      "put": {
        "tags": ["groups"],
        "summary": "Update Group Details",
        "description": "Update the details of an existing group.",
        "operationId": "updateGroupDetails",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/update_group_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group details updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/update_group_response"
                }
              }
            }
          },
          "404": { "description": "Group not found" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/{groupId}/join": {
      "post": {
        "tags": ["groups"],
        "summary": "Send Join Request to a Group",
        "description": "Allows an authenticated user to send a request to join a specific group.",
        "operationId": "sendJoinRequest",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Join request sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/join_request_response"
                }
              }
            }
          },
          "404": { "description": "Group not found" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/{groupId}/cancel": {
      "post": {
        "tags": ["groups"],
        "summary": "Cancel Join Request",
        "description": "Allows an authenticated user to cancel their pending join request for a specific group.",
        "operationId": "cancelJoinRequest",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Join request canceled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cancel_request_response"
                }
              }
            }
          },
          "404": { "description": "Group not found" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/{groupId}/members": {
      "get": {
        "tags": ["groups"],
        "summary": "Retrieve Group Members",
        "description": "Fetches a list of members for a specific group.",
        "operationId": "getGroupMembers",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of members in the specified group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_group_members_response"
                }
              }
            }
          },
          "404": { "description": "Group not found" },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/group/{groupId}/leave": {
      "delete": {
        "tags": ["groups"],
        "summary": "Leave a Group",
        "description": "Allows an authenticated user to leave a specific group they are a member of.",
        "operationId": "leaveGroup",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully left the group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/leave_group_response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User is not a member of the specified group."
          },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "500": { "description": "Internal server error." }
        }
      }
    },
    "/api/group/delete/{groupId}": {
      "delete": {
        "tags": ["groups"],
        "summary": "Delete a Group",
        "description": "Allows the creator of a group to delete it permanently.",
        "operationId": "deleteGroupByGroupId",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/delete_group_response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - The user is not the owner of the group."
          },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorization)."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "404": { "description": "Not Found - The group does not exist." },
          "500": { "description": "Internal server error." }
        }
      }
    },
    "/api/group/{groupId}/requests": {
      "get": {
        "tags": ["groups"],
        "summary": "Get Pending Join Group Requests",
        "description": "Allows the owner of a group to view all pending join requests for their group.",
        "operationId": "viewPendingRequests",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/view_pending_requests_response"
                  },
                  "example": [
                    {
                      "id": 1,
                      "email": "user@example.com",
                      "firstname": "John",
                      "lastname": "Doe",
                      "request_date": "2024-11-29T14:00:00Z"
                    },
                    {
                      "id": 2,
                      "email": "jane.doe@example.com",
                      "firstname": "Jane",
                      "lastname": "Doe",
                      "request_date": "2024-11-28T10:30:00Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - The user is not the owner of the group."
          },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorizatio."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "404": { "description": "Not Found - The group does not exist." },
          "500": { "description": "Internal server error." }
        }
      }
    },
    "/api/group/{groupId}/requests/{requestId}": {
      "put": {
        "tags": ["groups"],
        "summary": "Update Join Request Status",
        "description": "Allows the owner of a group to accept or reject a join request for their group.",
        "operationId": "updateJoinRequestStatus",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "The unique ID of the group.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "The unique ID of the join request.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The new status of the join request.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/update_join_requests_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Join request status updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/update_join_requests_response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - The user is not the owner or invalid status"
          },
          "401": {
            "description": "Unauthorized - Incorrect password or invalid authorizatio."
          },
          "403": { "description": "Forbidden - Invalid credentials provided." },
          "404": { "description": "Not Found - The group does not exist." },
          "500": { "description": "Internal server error." }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "group_response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the group"
          },
          "name": {
            "type": "string",
            "description": "Name of the group"
          },
          "description": {
            "type": "string",
            "description": "Description of the group"
          },
          "creator_id": {
            "type": "integer",
            "description": "ID of the group creator (can be null)"
          }
        },
        "example": {
          "id": 1,
          "name": "Developers Group",
          "description": "A group for developers",
          "creator_id": 101
        }
      },
      "group_detail_response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the group"
          },
          "name": {
            "type": "string",
            "description": "Name of the group"
          },
          "description": {
            "type": "string",
            "description": "Description of the group"
          },
          "creator_id": {
            "type": "integer",
            "description": "ID of the group creator (can be null)"
          },
          "member_count": {
            "type": "integer",
            "description": "Number of members in the group"
          }
        },
        "example": {
          "id": 1,
          "name": "Developers Group",
          "description": "A group for developers",
          "creator_id": 101,
          "member_count": 15
        }
      },
      "create_group_body": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the group",
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "description": "A brief description of the group",
            "maxLength": 200
          }
        },
        "example": {
          "name": "Developers Group",
          "description": "A group for developers to collaborate"
        }
      },
      "create_group_response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the group"
          },
          "name": {
            "type": "string",
            "description": "The name of the group"
          },
          "description": {
            "type": "string",
            "description": "A brief description of the group"
          },
          "creator_id": {
            "type": "integer",
            "description": "ID of the user who created the group"
          }
        },
        "example": {
          "id": 1,
          "name": "Developers Group",
          "description": "A group for developers to collaborate",
          "creator_id": 101
        }
      },
      "get_group_response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Group Name"
          },
          "description": {
            "type": "string",
            "example": "Group description"
          },
          "creator_id": {
            "type": "integer",
            "example": 10
          }
        },
        "example": {
          "id": 1,
          "name": "Study Group",
          "description": "A group for students to collaborate on homework and projects.",
          "creator_id": 10
        }
      },
      "update_group_body": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The updated name of the group",
            "example": "Updated Group Name"
          },
          "description": {
            "type": "string",
            "description": "The updated description of the group",
            "example": "Updated group description"
          }
        },
        "required": ["name", "description"]
      },
      "update_group_response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Group Name"
          },
          "description": {
            "type": "string",
            "example": "Group description"
          },
          "creator_id": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "join_request_response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Join request sent successfully"
          },
          "request": {
            "type": "object",
            "description": "Details of the join request",
            "properties": {
              "id": {
                "type": "integer",
                "example": 123
              },
              "group_id": {
                "type": "integer",
                "example": 10
              },
              "account_id": {
                "type": "integer",
                "example": 5
              },
              "status": {
                "type": "string",
                "example": "pending"
              },
              "request_date": {
                "type": "string",
                "format": "date-time",
                "example": "2024-11-29T12:34:56Z"
              },
              "response_date": {
                "type": "string",
                "format": "date-time",
                "example": "2024-11-30T12:34:56Z"
              }
            }
          }
        },
        "example": {
          "message": "Join request sent successfully",
          "request": {
            "id": 123,
            "group_id": 10,
            "account_id": 5,
            "status": "pending",
            "request_date": "2024-11-29T12:34:56Z",
            "response_date": null
          }
        }
      },
...
      "get_group_members_response": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "email": {
              "type": "string",
              "example": "user@example.com"
            },
            "firstname": {
              "type": "string",
              "example": "John"
            },
            "lastname": {
              "type": "string",
              "example": "Doe"
            }
          }
        },
        "example": [
          {
            "id": 1,
            "email": "user1@example.com",
            "firstname": "John",
            "lastname": "Doe"
          },
          {
            "id": 2,
            "email": "user2@example.com",
            "firstname": "Jane",
            "lastname": "Smith"
          }
        ]
      },
...
    }
  }
}